Issue 32 - Impossible to play some video files with Video.DirectShow
The issue was cause by fixing issue 22, which is committed in revision 382. Removing call of graph's Render() method solves the issue with freezing video after screen saver, but introduces another issue when playing some video files. The Render() method is adding not only renderer filter, but some other filters, which may be required. To configure this, a new property was introduced - PreventFreezing, which allows to select what is requires - freezing preventing removing Render() call or not preventing leaving it.


git-svn-id: http://aforge.googlecode.com/svn/trunk@418 ab5b2c17-5f1b-0410-93cd-b5c26ce635c7