Issue 228: Resolution of Bitmap returned by BaseTransformationFilter.Apply is default
Setting horizontal/vertical resolution of result image to the same values as source image has. The fix is applied to filters and image processing routines which accept Bitmap as parameter and return Bitmap as result. Obviously it can not copy DPI settings when image processing routine is applied to BitmapData or UnmanagedImage.

git-svn-id: http://aforge.googlecode.com/svn/trunk@1605 ab5b2c17-5f1b-0410-93cd-b5c26ce635c7