r1727, https://code.google.com/p/aforge/source/detail?r=1727
Fixed Erosion and Dilation filters so they keep pixel as is when
processing image borders and all relevant values of a structuring
element (SE) are outside of image at the current SE position.

For example, the below SE would reveal the bug when left image border is
processed. For all pixels on an image's left border, there are no pixels
to process for MIN/MAX operator, so pixels should be kept unchanged.
1 -1 -1
1 -1 -1
1 -1 -1